# Remove unbounded cache