`TUFMetaStorage` uses an in-memory cache to store all retrieved metadata files. This cache has a few problems:

1. It doesn't have any sort of limit on its size so it will grow indefinitely as more content is retrieved
2. Entries in the cache are not deleted when `Delete` is calls on a `TUFMetaStorage` instance. This means the cache can hold entries that have been deleted from the underlying storage (there seem to be tests for this behavior so could this be for a reason?)
3. It isn't safe for concurrent use, see #1613.

I've taken the scorched-earth approach of simply removing this cache. After #1639 it should be fast to get these files from the database each time. I'll test this on Docker's production instance before merging it to check the performance is still OK.