@jonnystoten 2022-09-16
Looks good on Docker's notary server after a few days. The performance seems fine and more consistent - the average is a little bit slower (~3ms -> ~5ms) but the worst case is much better (>100ms -> ~8ms).

I'm assuming the tests that were testing that deleting from the TUF store doesn't delete from the cache were just wrong - anything relying on this will have other issues because of course if the server is restarted the cache will be cleared and if more than one server is running each one will have different cache contents etc.

This can also close #1613.